double -> int (error correction)

+++ src/main/java/com/innotec/bats/general/Account.java
@@ -11,7 +11,7 @@ public class Account implements java.io.Serializable
 	private String accountHolderIdNo;
 	public static final double DEFAULT_MAX_WITHDRAWAL_PER_DAY = 1000.00;
 	public static final double DEFAULT_MAX_TRANSFER_PER_DAY = 1000.00;
-	public static final double ACCOUNT_NO_LEN = 13;
+	public static final int ACCOUNT_NO_LEN = 13;
 	public Account (String accountNo, double balance, boolean active,
 			double maxWithdrawalPerDay, double maxTransferPerDay,
 			String accountHolderIdNo)
